requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
Django==3.1.2
djangorestframework==3.12.1
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.9.27
requests==2.24.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
zipp==3.3.0
